Feat: separating codes related to implementation of linked list from mutex module.c
